+++ storm-core/src/jvm/backtype/storm/messaging/netty/SaslMessageToken.java
@@ -17,14 +17,23 @@
  */
 package backtype.storm.messaging.netty;
 
+import java.io.IOException;
 import org.jboss.netty.buffer.ChannelBuffer;
 import org.jboss.netty.buffer.ChannelBufferOutputStream;
 import org.jboss.netty.buffer.ChannelBuffers;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
  * Send and receive SASL tokens.
  */
-public class SaslMessageToken {
+public class SaslMessageToken implements INettySerializable {
+    public static final short IDENTIFIER = -500;
+
+    /** Class logger */
+    private static final Logger LOG = LoggerFactory
+            .getLogger(SaslMessageToken.class);
+
     /** Used for client or server's token to send or receive from each other. */
     private byte[] token;
 
@@ -63,31 +72,43 @@ public class SaslMessageToken {
         this.token = token;
     }
 
-    int encodeLength() {
+    public int encodeLength() {
         return 2 + 4 + token.length;
     }
 
     /**
      * encode the current SaslToken Message into a channel buffer
-     * SaslTokenMessageRequest is encoded as: identifier .... short(2) always it
-     * is -500 payload length .... int payload .... byte[]
+     * SaslTokenMessageRequest is encoded as: identifier .... short(2)
+     * payload length .... int payload .... byte[]
      * 
-     * @throws Exception
+     * @throws IOException
      */
-    ChannelBuffer buffer() throws Exception {
+    public ChannelBuffer buffer() throws IOException {
         ChannelBufferOutputStream bout = new ChannelBufferOutputStream(
                 ChannelBuffers.directBuffer(encodeLength()));
-        short identifier = -500;
         int payload_len = 0;
         if (token != null)
             payload_len = token.length;
 
-        bout.writeShort(identifier);
+        bout.writeShort(IDENTIFIER);
         bout.writeInt(payload_len);
+
         if (payload_len > 0) {
             bout.write(token);
         }
         bout.close();
         return bout.buffer();
+    }
+    
+    public static SaslMessageToken read(byte[] serial) {
+        ChannelBuffer sm_buffer = ChannelBuffers.copiedBuffer(serial);
+        short identifier = sm_buffer.readShort();
+        int payload_len = sm_buffer.readInt();
+        if(identifier != IDENTIFIER) {
+            return null;
+        }
+        byte token[] = new byte[payload_len];
+        sm_buffer.readBytes(token, 0, payload_len);
+        return new SaslMessageToken(token);
     }
 }