+++ storm-core/src/jvm/backtype/storm/messaging/netty/StormServerHandler.java
@@ -17,7 +17,8 @@
  */
 package backtype.storm.messaging.netty;
 
-import backtype.storm.messaging.TaskMessage;
+import backtype.storm.utils.Utils;
+import org.jboss.netty.channel.Channel;
 import org.jboss.netty.channel.ChannelHandlerContext;
 import org.jboss.netty.channel.ChannelStateEvent;
 import org.jboss.netty.channel.ExceptionEvent;
@@ -25,33 +26,39 @@ import org.jboss.netty.channel.MessageEvent;
 import org.jboss.netty.channel.SimpleChannelUpstreamHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
 
-class StormServerHandler extends SimpleChannelUpstreamHandler  {
+public class StormServerHandler extends SimpleChannelUpstreamHandler  {
     private static final Logger LOG = LoggerFactory.getLogger(StormServerHandler.class);
-    Server server;
+    IServer server;
     private AtomicInteger failure_count; 
+    private Channel channel;
     
-    StormServerHandler(Server server) {
+    public StormServerHandler(IServer server) {
         this.server = server;
         failure_count = new AtomicInteger(0);
     }
     
     @Override
     public void channelConnected(ChannelHandlerContext ctx, ChannelStateEvent e) {
-        server.addChannel(e.getChannel());
+        server.channelConnected(e.getChannel());
+        if(channel != null) {
+            LOG.debug("Replacing channel with new channel: {} -> ",
+                      channel, e.getChannel());
+        }
+        channel = e.getChannel();
+        server.channelConnected(channel);
     }
     
     @Override
     public void messageReceived(ChannelHandlerContext ctx, MessageEvent e) {
-      List<TaskMessage> msgs = (List<TaskMessage>) e.getMessage();
+      Object msgs = e.getMessage();
       if (msgs == null) {
         return;
       }
       
       try {
-        server.enqueue(msgs, e.getRemoteAddress().toString());
+        server.received(msgs, e.getRemoteAddress().toString(), channel);
       } catch (InterruptedException e1) {
         LOG.info("failed to enqueue a request message", e);
         failure_count.incrementAndGet();
@@ -61,6 +68,7 @@ class StormServerHandler extends SimpleChannelUpstreamHandler  {
     @Override
     public void exceptionCaught(ChannelHandlerContext ctx, ExceptionEvent e) {
         LOG.error("server errors in handling the request", e.getCause());
+        Utils.handleUncaughtException(e.getCause());
         server.closeChannel(e.getChannel());
     }
 }