+++ storm-core/src/jvm/backtype/storm/metric/internal/LatencyStatAndMetric.java
@@ -19,11 +19,10 @@ package backtype.storm.metric.internal;
 
 import java.util.Map;
 import java.util.HashMap;
-import java.util.Timer;
 import java.util.TimerTask;
-import java.util.concurrent.atomic.AtomicLong;
 
 import backtype.storm.metric.api.IMetric;
+import backtype.storm.utils.Utils;
 
 /**
  * Acts as a Latency Metric, but also keeps track of approximate latency
@@ -145,7 +144,9 @@ public class LatencyStatAndMetric implements IMetric {
         }
 
         long timeSpent = now - _bucketStart;
-        double ret = ((double)(lat + _exactExtraLat))/(count + _exactExtraCount);
+        long exactExtraCountSum = count + _exactExtraCount;
+        double ret = Utils.zeroIfNaNOrInf(
+                ((double) (lat + _exactExtraLat)) / exactExtraCountSum);
         _bucketStart = now;
         _exactExtraLat = 0;
         _exactExtraCount = 0;
@@ -227,7 +228,9 @@ public class LatencyStatAndMetric implements IMetric {
         ret.put("600", readApproximateLatAvg(lat, count, timeSpent, _tmTime, _tmLatBuckets, _tmCountBuckets, 600 * 1000));
         ret.put("10800", readApproximateLatAvg(lat, count, timeSpent, _thTime, _thLatBuckets, _thCountBuckets, 10800 * 1000));
         ret.put("86400", readApproximateLatAvg(lat, count, timeSpent, _odTime, _odLatBuckets, _odCountBuckets, 86400 * 1000));
-        ret.put(":all-time", ((double)lat + _allTimeLat)/(count + _allTimeCount));
+        long allTimeCountSum = count + _allTimeCount;
+        ret.put(":all-time", Utils.zeroIfNaNOrInf(
+                (double) lat + _allTimeLat)/allTimeCountSum);
         return ret;
     }
 
@@ -242,7 +245,7 @@ public class LatencyStatAndMetric implements IMetric {
             totalCount += countBuckets[i];
             timeNeeded -= bucketTime[i];
         }
-        return ((double)totalLat)/totalCount;
+        return Utils.zeroIfNaNOrInf(((double) totalLat) / totalCount);
     }
 
     public void close() {